Fixed resizing bug
Right now everytime the cursor retrieves a pixel position a ratio is applied so the calculations in the image never get out of limits.
This is necessary because the QPixmap is being resized and the cursor position is given according to that resized QPixmap but the QImage
where the data exists is never resized.